Bug fixes and experiments
Developed Scripts for model development - created 12 models

Bug Fix in validation to make sure files are saved to the correct location

Update to the functions that automatically detect which vegetation indices should be used.  Now if the system is unable to determine the correct method of determining the veg indices, it will simply calculate them all.  This may be significantly slower, but in my opinion follows the principle of least surprise.